+++ SurfEngine/src/SurfEngine/Platform/OpenGl/OpenGLVertexArray.cpp
@@ -56,7 +56,8 @@ namespace SurfEngine {
 				ShaderDataTypeToOpenGLBaseType(element.Type),
 				element.Normalized ? GL_TRUE : GL_FALSE,
 				vertexBuffer->GetLayout().GetStride(),
-				(const void*)element.Offset);
+				(const void*)(uint64_t)element.Offset
+			);
 			index++;
 		}
 		m_VertexBuffers.push_back(vertexBuffer);
